✨ feat: Update database backup and restore scripts
This commit updates the database backup and restore scripts to include additional flags for `pg_dumpall` command, specifically "--no-owner" and "--no-acl", to avoid issues when restoring on different environments. It also renames the "OUTPUT_DIR" variable in the `postgres.sh` script to "BACKUP_DIR" to avoid confusion with the output directory of the `backup.sh` script.

It also removes references to unused variables (`_TMP_PATH`, `_STATE_PATH`, `_STOWED_PATH`) in `data.sh` and `backup.sh`.

Finally, it includes a new script `restore.sh`, which restores a database backup to the specified namespace. It takes two optional arguments: `-n` to specify the namespace and `-b` to specify the backup directory. If no backup directory is specified, it exits with an error message.